pyproject.toml
[project]
name = "smbv2"
version = "0.1.0"
description = "Document Ingestor"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pdfplumber>=0.11.6",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pymupdf>=1.26.0",
    "pytesseract>=0.3.13",
    "sentence-transformers>=4.1.0",
    "transformers>=4.52.4",
]
[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
[tool.uv]
python-install-mirror = "https://mirror.nju.edu.cn/github-release/indygreg/python-build-standalone"
